% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cknneighbors_graph.R
\name{cknneighbors_graph}
\alias{cknneighbors_graph}
\title{Implementation of of Continuous k-Nearest Neighbors (CKNNs) in R}
\usage{
cknneighbors_graph(k = 2, lambda = 1, edges)
}
\arguments{
\item{k}{integer: number of neighbors}

\item{lambda}{double: positive parameter regulating density.}

\item{edges}{dataframe: weighted undirected edge list for similarity between
trajectories.}
}
\value{
A dataframe of edges with CKNNs selected edges only.
}
\description{
CKNNs connect points x,y if \eqn{d(x,y) < lambda * (d(x,x_k) * d(y,y_k))^0.5}, where \eqn{d(x,y)} is the distance between points x and y, and variables \eqn{x_k} and \eqn{y_k}
are the k-th closest neighbors of points x and y, respectively. See \url{https://arxiv.org/pdf/1606.02353.pdf} for details.
}
\author{
Ashleigh C. Myall (\email{a.myall19@imperial.ac.uk})
}
