% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_trajectories.R
\name{example_trajectories}
\alias{example_trajectories}
\title{Generate an example dataset}
\usage{
example_trajectories()
}
\value{
Dataframe in long format with each row contiang the unique location-time
positions of each trajectory
}
\description{
A pre-built function with no arguements to return example trajectories. Each trajectory
contains a location (ward), and a corresponding time (date) on that location.
}
\author{
Ashleigh C. Myall (\email{a.myall19@imperial.ac.uk})
}
