% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialTempProx.R
\name{getSpatialTempProx}
\alias{getSpatialTempProx}
\title{Compute full space and time between all trajectories}
\usage{
getSpatialTempProx(traj.l, D, beta = 0.6)
}
\arguments{
\item{traj.l}{list: list of all trajectories}

\item{D}{matrix: effective distance matrix, with elements the shortest path effective
distance weight between nodes i and j}

\item{beta}{double: paramter regulating the effect of time in spatial-temporal
proximity. beta can also be intepreted as speed of proporgation across the
background mobility data. To pass when for the spatial temporal proximity kernal (matrix computation).}
}
\value{
A weighted-edge list of similarities between trajectories.
}
\description{
Given the total set of trajectories, compute proxmity between all combinations.
}
\author{
Ashleigh C. Myall (\email{a.myall19@imperial.ac.uk})
}
