% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernal_dist.R
\name{kernal_dist}
\alias{kernal_dist}
\title{Spatial temporal proximity kernal (matrix computation)}
\usage{
kernal_dist(space_d, time_d, beta = 0)
}
\arguments{
\item{space_d}{matrix: pairwise space (network) distances between locations of
two trajectories.}

\item{time_d}{matrix: pairwise time distances between locations of
two trajectories.}

\item{beta}{double: paramter regulating the effect of time in spatial-temporal
proximity. beta can also be intepreted as speed of proporgation across the
background mobility data.}
}
\value{
Sum of spatial-temporal proximities
}
\description{
Computes total spatial temporal proximities using the kernal.
}
\section{Details}{

For every pair of patients, we define spatio-temporal proximity between ward-time locations
\eqn{l_{i}} and \eqn{l_{j}} with the kernel:

\deqn{\kappa(l_{i},l_{j})  = e^{-\delta_{ij}-\beta \tau_{ij}}}

where \eqn{tau_{ij}=\begin{vmatrix}t_i - t_j\end{vmatrix}}, parameter \eqn{\beta}
represents a propagation speed, and \eqn{\delta_{ij}} denotes the shortest-path
distance (the most probable pathway for disease propagation) between wards
\eqn{v_i} and \eqn{v_j} across the background movement network \eqn{G}.
Note the equation reaches a maximum of one when
\eqn{l_{ia} = l_{jb}} (exact overlaps), and decays to zero as spatial-temporal
proximity becomes more distant. We then measure overall similarity between
trajectories \eqn{T_m} and \eqn{T_n} by summing over pairwise proximity measures
between \eqn{l_i \in T_n$ and $l_j \in T_m}:

\deqn{S(T_n,T_m) =  \sum_{l_i \in T_n}\sum_{l_j \in T_m}\kappa(l_{i},l_{j})}

A full description can be found in our paper: \url{https://doi.org/10.1101/2021.04.07.21254497}
about this package.
}

\author{
Ashleigh C. Myall (\email{a.myall19@imperial.ac.uk})
}
