% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cknneighbors_graph.R
\name{cknneighbors_graph}
\alias{cknneighbors_graph}
\title{Implementation of of Continuous k-Nearest Neighbors (CKNN) in R}
\usage{
cknneighbors_graph(k = 2, lambda = 1, edges)
}
\arguments{
\item{k}{integer: number of neighbors}

\item{lambda}{double: positive parameter regulating density.}

\item{edges}{dataframe: weighted undirected edge list for similarity between
trajectories.}
}
\value{
dataframe of edges with CKNN selected edges only.
}
\description{
See https://arxiv.org/pdf/1606.02353.pdf. CKNN connect points x,y

if d(x,y) < lambda * ( d(x,x_k) * d(y,y_k) )^0.5

Where d(x,y) is the distance between points x and y. And x_k, and y_k, are the
k-th closest neighbors of points x and y respectively.
}
\author{
Ashleigh C. Myall (\email{a.myall19@imperial.ac.uk})
}
