% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot pathways function}
\usage{
plot_trajectories(traj.l)
}
\arguments{
\item{traj.l}{List: list of trajectory dataframes}
}
\value{
GGplot visulisation
}
\description{
Build ggplot visualisation for time components of trajectories
}
\author{
Ashleigh C. Myall (\email{a.myall19@imperial.ac.uk})
}
